+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.1.9-dev.1
 
+-   rename Connection.sendRequest param modelName
+
 ## 0.1.8-dev.1
 
 -   Support for RPC commands.